fix teacher comment bug and add valid constraints on post and put request

+++ src/main/java/com/rmr101/campus/entity/StudentAssignment.java
@@ -18,7 +18,7 @@ public class StudentAssignment {
     private String attachmentUrl;
 
     private boolean isScored;
-    private float score;
+    private int score;
     private String comment;
 
     @ManyToOne(fetch=FetchType.LAZY)
